requirements.txt
absl-py==1.0.0
brotlipy==0.7.0
cachetools==5.0.0
certifi==2022.9.24
cffi==1.15.0
charset-normalizer==2.0.4
cloudpickle==2.2.0
contextlib2==21.6.0
cryptography==36.0.0
cycler==0.11.0
cytoolz==0.11.0
dask==2022.2.0
decorator==5.1.1
fonttools==4.37.2
fsspec==2022.8.2
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.44.0
idna==3.3
imagecorruptions==1.1.2
imageio==2.22.0
imgaug==0.4.0
importlib-metadata==4.11.3
joblib==1.1.0
kiwisolver==1.4.4
kornia==0.6.7
locket==1.0.0
Markdown==3.3.6
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
ml-collections==0.1.1
networkx==2.5
numpy==1.21.2
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
partd==1.3.0
Pillow==9.0.1
pip==21.2.2
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
PyWavelets==1.1.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.16.2
scikit-learn==1.0.2
scipy==1.7.3
setuptools==58.0.4
Shapely==1.7.1
six==1.16.0
sklearn==0.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
toolz==0.12.0
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.63.1
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.1.1
wheel==0.37.1
zipp==3.8.0
